order_transfer/views.py
from django.shortcuts import render

from order_transfer.models import Order
from order_transfer.serializers import OrderSerializer
from rest_framework import viewsets

class OrderViewSet(viewsets.ModelViewSet):
    queryset = Order.objects.all()
    serializer_class = OrderSerializer